Update schema for application status history
Creates a new join table so that we can keep track of prior
statuses of an application, not just the most recent status.
This will allow clients to display the history of changes to
an application's status, or do analysis on things like how long
a company or position takes to move candidates through the pipeline.

In addition to creating the new table `application_status_changes`:

- refactor `applications` to drop the `application_status_id` column
- drop the `date_applied` column from `applications` (this would
duplicate information from `application_status_changes`)
- update the `applications_joined` view to return the most recent
status of an application and when that status was set
- added some clarifying comments in schema and seed files