Include reference to user who posted job
This unlocks future implementations of auth / permissions
where a job can be edited or deleted only by the person who
added it (and an admin).

This involved adding a foreign key to `users` in the `jobs` table,
as well as updates to the `jobs_joined` view and the seeds file.